+++ packages/ulangi-mobile/src/views/quiz/QuizWritingScreenContainer.tsx
@@ -57,7 +57,7 @@ export class QuizWritingScreenContainer extends Container<
     new ObservableWritingFormState(
       true,
       null,
-      this.questionIterator.next(),
+      this.questionIterator.current(),
       '',
       '',
       0,